Change domain go creekservice
https://github.com/creek-service/creek-test/issues/24
