+++ conformity/src/main/java/module-info.java
@@ -1,6 +1,6 @@
 module creek.test.conformity {
     requires transitive creek.base.annotation;
 
-    exports org.creek.api.test.conformity;
-    exports org.creek.api.test.conformity.check;
+    exports org.creekservice.api.test.conformity;
+    exports org.creekservice.api.test.conformity.check;
 }
