+++ hamcrest/src/main/java/module-info.java
@@ -2,5 +2,5 @@ module creek.test.hamcrest {
     requires transitive org.hamcrest;
     requires transitive com.github.spotbugs.annotations;
 
-    exports org.creek.api.test.hamcrest;
+    exports org.creekservice.api.test.hamcrest;
 }
