+++ util/src/main/java/module-info.java
@@ -3,6 +3,6 @@ module creek.test.util {
     requires transitive com.github.spotbugs.annotations;
     requires transitive java.management;
 
-    exports org.creek.api.test.util;
-    exports org.creek.api.test.util.coverage;
+    exports org.creekservice.api.test.util;
+    exports org.creekservice.api.test.util.coverage;
 }
